+++ packages/happner-2/test/integration/security/user-group-management-mongo.js
@@ -298,7 +298,7 @@ require('../../__fixtures/utils/test_helper').describe({ timeout: 120e3 }, (test
 
               testUserClient.exchange.security.updateOwnUser(testUser, function (e) {
                 if (!e) return done(new Error('error was test.expected'));
-                test.expect(e.toString()).to.be('Error: missing oldPassword parameter');
+                test.expect(e.toString()).to.be('Error: Bad Password Arguments');
                 done();
               });
             })
@@ -361,87 +361,13 @@ require('../../__fixtures/utils/test_helper').describe({ timeout: 120e3 }, (test
 
               testUserClient.exchange.security.updateOwnUser(testUser, function (e) {
                 if (!e) return done(new Error('error was test.expected'));
-                test.expect(e.toString()).to.be('Error: old password incorrect');
+                test.expect(e.toString()).to.be('SystemError: Invalid old password');
                 done();
               });
             })
             .catch(function (e) {
               done(e);
             });
-        });
-      });
-    });
-  });
-
-  it('adds a test user, logs in with the test user - modifies the user details without the users password changing', function (done) {
-    var testGroup = {
-      name: 'TESTGROUP6' + test_id,
-
-      custom_data: {
-        customString: 'custom1',
-        customNumber: 0,
-      },
-
-      permissions: {
-        methods: {},
-      },
-    };
-
-    var testGroupSaved;
-    var testUserSaved;
-    var testUserClient;
-
-    adminClient.exchange.security.addGroup(testGroup, function (e, result) {
-      if (e) return done(e);
-
-      testGroupSaved = result;
-
-      var testUser = {
-        username: 'TESTUSER6' + test_id,
-        password: 'TEST PWD',
-        custom_data: {
-          something: 'useful',
-        },
-        application_data: {
-          something: 'sacred',
-        },
-      };
-
-      adminClient.exchange.security.addUser(testUser, function (e, result) {
-        if (e) return done(e);
-
-        test.expect(result.username).to.be(testUser.username);
-        testUserSaved = result;
-        adminClient.exchange.security.linkGroup(testGroupSaved, testUserSaved, function (e) {
-          if (e) return done(e);
-          testUserClient = new Mesh.MeshClient({ secure: true, port: 8003 });
-
-          testUserClient
-            .login(testUser)
-            .then(function () {
-              delete testUser.password;
-              testUser.custom_data = { changedCustom: 'changedCustom' };
-              testUser.application_data = { something: 'profane' };
-              return testUserClient.exchange.security.updateOwnUser(testUser);
-            })
-            .then(function (result) {
-              test.expect(result.custom_data.changedCustom).to.be('changedCustom');
-              test.expect(result.application_data.something).to.be('sacred');
-              return adminClient.exchange.security.getUser(testUser.username);
-            })
-            .then(function (result) {
-              test.expect(result.application_data.something).to.be('sacred');
-              testUser.application_data = { something: 'profane' };
-              return adminClient.exchange.security.updateUser(testUser);
-            })
-            .then(function () {
-              return adminClient.exchange.security.getUser(testUser.username);
-            })
-            .then(function (result) {
-              test.expect(result.application_data.something).to.be('profane');
-              done();
-            })
-            .catch(done);
         });
       });
     });