+++ packages/happn-cluster/RELEASES.md
@@ -359,3 +359,8 @@ non-release 2017-12-14
 -----------------
   - feat: changePassword in client
   - feat: resetPassword supported by auth provider
+
+12.4.1 2023-05-24
+-----------------
+  - fix: happn-3 resetpassword
+  - feat: happner-2 resetpassword