+++ packages/happner-client/RELEASES.md
@@ -423,3 +423,8 @@
 -----------------
   - feat: happn-3 changePassword in client
   - feat: happn-3 resetPassword supported by auth provider
+
+12.7.1 2023-05-24
+-----------------
+  - fix: happn-3 resetpassword
+  - feat: happner-2 resetpassword